refactor(core): small reorganization
