+++ src/core/hooks/use-formts/index.ts
@@ -1,0 +1,1 @@
+export { useFormts, FormtsOptions } from "./use-formts";
